requirements.txt
# CoDA Requirements
# Base requirements for CoDA project
# MMaDA-specific requirements are in MMaDA/requirements.txt

# Core dependencies
torch>=2.0.0
numpy>=1.21.0
pillow>=9.0.0

# Configuration and utilities
omegaconf>=2.3.0
pyyaml>=6.0

# Evaluation and metrics
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Logging and experiment tracking
wandb>=0.15.0
tensorboard>=2.10.0

# Data handling
pandas>=1.3.0
h5py>=3.7.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

